requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
httpx==0.27.0
pytest==8.1.1
groq==0.5.0
python-dotenv==1.0.1
requests==2.31.0
